requirements.txt
streamlit>=1.24.0
python-dotenv
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.1
sentencepiece>=0.1.99
protobuf>=4.25.1
